Change public imports to common (#375)
